Adding the service implementation. See cbd34081517f723917afb37e1761a68aaa695151
If a new bird count is started through an activity, the BirdCount object would have to be passed to every sub-activity. Thus the BirdCount class and each subclass would have to implement the Parcelable interface for reasonable performance. By using a separate service we circumvent this and provide an elegant solution for accessing the current bird count as well as adding observations.

To create a service we could either

* a bound service, or
* a started service

What we want is to connect to the service through each activity. It could then modify the current bird count as it needs to. Thus using a bound service seems OK. However we face some implementation problems, as a bound service will automatically destroyed when no more activities are connected to it and each activity should disconnect not later than at Activity#onDestroy(). Thus the last connections may already be closed before the newly started activity was able to connect and the service might be closed To prevent this from happening, we use a started service: the first activity which may need the service will start it if necessary and the service will automatically stop itself when the census is finished. All activities will connect to the service the usual way.